+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.1](https://github.com/cssninjaStudio/nuxt-toaster/compare/v0.1.0...v0.1.1) (2022-09-25)
+
+
+### Features
+
+* allow class array in containerClass and wrapperClass theme ([9296daf](https://github.com/cssninjaStudio/nuxt-toaster/commit/9296daf966f9652e4490d249159217c31c38d1f2))
+
+## [0.1.0](https://github.com/cssninjaStudio/nuxt-toaster/compare/v0.0.3...v0.1.0) (2022-09-24)
+
+
+### ⚠ BREAKING CHANGES
+
+* move base options from nuxt config to app config
+
+### Features
+
+* move base options from nuxt config to app config ([4145b25](https://github.com/cssninjaStudio/nuxt-toaster/commit/4145b25ecd0da337777f21e10931385a9fa9ce8d))
+
+
+### Bug Fixes
+
+* define props override issue ([c8097d4](https://github.com/cssninjaStudio/nuxt-toaster/commit/c8097d4cc8f8ab57ffa0af81bfb74f401c47b0b7))
+
+### [0.0.3](https://github.com/cssninjaStudio/nuxt-toaster/compare/v0.0.2...v0.0.3) (2022-09-23)
+
+
+### Bug Fixes
+
+* queues and events propagation ([3aaab8f](https://github.com/cssninjaStudio/nuxt-toaster/commit/3aaab8f30d5def71b05c968f43e5d8802699ea2c))
+
 ### [0.0.2](https://github.com/cssninjaStudio/nuxt-toaster/compare/v0.0.1...v0.0.2) (2022-09-23)